requirements.txt
flask==2.2.2  # Web framework
tensorflow==2.9.1  # For loading .h5 model
numpy==1.23.1  # Numerical computations
pandas==1.4.3  # Data handling (if needed)
opencv-python==4.5.5.62  # Image processing 
pillow==9.2.0  # Image handling in Python
